+++ tests/lts_compatibility.py
@@ -5,11 +5,9 @@ import infra.e2e_args
 import infra.proc
 import infra.logging_app as app
 import infra.utils
+import infra.gh_helper
 import os
-import re
-from github import Github
 
-from setuptools.extern.packaging.version import Version  # type: ignore
 from loguru import logger as LOG
 
 
@@ -83,7 +81,7 @@ def run_live_compatibility_since_last(args, lts_major_version, lts_install_path)
 
         try:
             network.wait_for_new_primary(new_primary)
-            assert False, "No new primary should be elected"
+            assert False, "No new primary should be elected while"
         except TimeoutError:
             pass
 
@@ -114,101 +112,6 @@ def run_ledger_compatibility_since_first(args, lts_releases):
     pass
 
 
-# This assumes that CCF is installed at `/opt/ccf`, which is true from 1.0.0
-INSTALL_DIRECTORY_SUB_PATH = "opt/ccf"
-BRANCH_RELEASE_PREFIX = "release/"
-REPOSITORY_NAME = "microsoft/CCF"
-DEBIAN_PACKAGE_EXTENSION = "_amd64.deb"
-
-
-def get_release_branches_names(repo):
-    return [
-        branch.name
-        for branch in repo.get_branches()
-        if branch.name.startswith(BRANCH_RELEASE_PREFIX)
-    ]
-
-
-def get_releases_from_release_branch(repo, branch_name):
-    # Assumes that N.a.b releases can only be cut from N.x branch, with N a valid major version number
-    assert branch_name.startswith(
-        BRANCH_RELEASE_PREFIX
-    ), f"{branch_name} is not a release branch"
-
-    release_branch_name = branch_name[len(BRANCH_RELEASE_PREFIX) :]
-    release_re = "^ccf-{}$".format(release_branch_name.replace(".x", "([.\d+]+)"))
-
-    # Most recent tag is first
-    return list(([tag for tag in repo.get_tags() if re.match(release_re, tag.name)]))
-
-
-def install_ccf_debian_package(debian_package_url, directory_name):
-    LOG.info(f"Downloading {debian_package_url}...")
-    download_cmd = ["wget", debian_package_url]
-    assert (
-        infra.proc.ccall(*download_cmd, log_output=False).returncode == 0
-    ), "Download failed"
-
-    LOG.info("Unpacking debian package...")
-    remove_cmd = ["rm", "-rf", directory_name]
-    assert (
-        infra.proc.ccall(*remove_cmd).returncode == 0
-    ), "Previous install cleanup failed"
-    install_cmd = ["dpkg-deb", "-R", debian_package_url.split("/")[-1], directory_name]
-    assert infra.proc.ccall(*install_cmd).returncode == 0, "Installation failed"
-
-    install_path = os.path.abspath(
-        os.path.join(directory_name, INSTALL_DIRECTORY_SUB_PATH)
-    )
-    LOG.success(f"CCF release successfully installed at {install_path}")
-    return install_path
-
-
-def install_latest_lts(args):
-    g = Github()
-    repo = g.get_repo(REPOSITORY_NAME)
-
-    with open(args.previous_lts_file) as f:
-        latest_release = f.readline()
-    latest_release_branch = f"release/{latest_release}"
-    # latest_release_branch = "release/0.99.x"  # TODO: To deduce from local checkout
-    LOG.info(f"Latest release branch for this checkout: {latest_release_branch}")
-
-    if latest_release_branch not in get_release_branches_names(repo):
-        raise ValueError(
-            f"Latest release branch {latest_release_branch} is not a valid release branch"
-        )
-
-    tags_for_this_release = get_releases_from_release_branch(
-        repo, latest_release_branch
-    )
-    LOG.info(f"Found tags: {[t.name for t in tags_for_this_release]}")
-
-    latest_tag_for_this_release = tags_for_this_release[0]
-    LOG.info(f"Most recent tag: {latest_tag_for_this_release.name}")
-
-    releases = [
-        r for r in repo.get_releases() if r.tag_name == latest_tag_for_this_release.name
-    ]
-    assert (
-        len(releases) == 1
-    ), f"Found {len(releases)} releases for tag {latest_tag_for_this_release.name}, expected 1"
-    release = releases[0]
-    LOG.info(f"Found release: {release.html_url}")
-
-    stripped_tag = latest_tag_for_this_release.name[len("ccf-") :]
-    debian_package_url = [
-        a.browser_download_url
-        for a in release.get_assets()
-        if re.match(f"ccf_{stripped_tag}{DEBIAN_PACKAGE_EXTENSION}", a.name)
-    ][0]
-
-    return stripped_tag, install_ccf_debian_package(
-        debian_package_url,
-        directory_name=f"ccf_install_{latest_tag_for_this_release.name}",
-    )
-
-
 if __name__ == "__main__":
 
     def add(parser):
@@ -224,14 +127,12 @@ if __name__ == "__main__":
     args.package = "libjs_generic"
     args.nodes = infra.e2e_args.max_nodes(args, f=0)
 
-    # TODO: Hardcoded because host only accepts from info on release builds
+    # Hardcoded because host only accepts from info on release builds
     args.host_log_level = "info"
 
-    lts_version, lts_install_path = install_latest_lts(args)
+    lts_major_version, lts_install_path = infra.gh_helper.install_latest_lts(args)
 
-    LOG.error(f"LTS version: {Version(lts_version).release[0]}")
+    LOG.error(f"LTS version: {lts_major_version}")
 
-    run_live_compatibility_since_last(
-        args, Version(lts_version).release[0], lts_install_path
-    )
+    run_live_compatibility_since_last(args, lts_major_version, lts_install_path)
     # run_ledger_compatibility_since_first(args, lts_releases)
